requirements.txt
langchain==0.3.23
langchain_chroma==0.2.2
langchain_community==0.3.21
langchain_core==0.3.51
langchain_huggingface==0.1.2
langchain_ollama==0.3.0
langchain-openai==0.3.12
langgraph==0.3.27
langgraph-checkpoint==2.0.24
langgraph-checkpoint-sqlite==2.0.6
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.61
python-dotenv==1.1.0
streamlit==1.44.1
openai==1.70.0
transformers==4.51.0
accelerate==1.6.0
sentence-transformers==4.0.2
unstructured==0.17.2
unstructured[md]
unstructured[pdf]